requirements.txt
pytest~=8.3.4
pytest-asyncio==0.25.3
pytest-playwright==0.5.1
playwright==1.45.1
httpx==0.24.1
allure-pytest==2.13.5
allure-python-commons~=2.13.5